+++ frontend/src/Components/responseBubble.jsx
@@ -28,72 +28,6 @@ ResponseBubble.propTypes = {
  */
 
 export default function ResponseBubble({ text }) {
-	const markdownText = `
-# 🧬 CRISPR-Cas9 in Human Gene Editing
-
-CRISPR-Cas9 is **revolutionizing modern medicine**, especially in the field of **gene editing**. Here's a breakdown of its impact and recent advancements:
-
----
-
-## 🧪 What is CRISPR-Cas9?
-
-CRISPR-Cas9 is a _genome editing tool_ derived from a bacterial defense system. It enables scientists to:
-
-- Cut DNA at specific locations  
-- Disable malfunctioning genes  
-- Insert or repair genes with high precision  
-
-> "It’s like a pair of molecular scissors guided by GPS." — Geneticist
-
----
-
-## ⚡ Recent Breakthroughs (2024–2025)
-
-### 1. **Sickle Cell Disease Therapy**
-- 🔬 *Technique:* Ex vivo CRISPR correction of patient hematopoietic stem cells  
-- ✅ *Result:* >90% of patients symptom-free in clinical trials
-
-### 2. **Cancer Immunotherapy**
-- 🧫 *Approach:* CRISPR-edited CAR-T cells  
-- 🎯 *Target:* Tumor-specific antigens  
-- 📈 *Outcome:* Improved T-cell persistence and tumor regression
-
-### 3. **Vision Restoration**
-- 👁 *Condition:* Leber Congenital Amaurosis (LCA)  
-- 🧬 *Tool:* In vivo editing using viral delivery  
-- 🌟 *Impact:* Partial sight restoration in blind patients
-
----
-
-## 📚 Code Example (just for fun!)
-
-\`\`\`python
-def edit_gene(sequence, target, replacement):
-    if target in sequence:
-        return sequence.replace(target, replacement)
-    return sequence
-
-print(edit_gene("ATCGGATC", "GGA", "TTT"))  # Simulated gene edit
-\`\`\`
-
----
-
-## 📌 Ethical Considerations
-
-- ⚖️ Germline editing remains controversial  
-- 🧒 Heritable changes could affect future generations  
-- 🧑‍⚖️ Global consensus still evolving
-
----
-
-## 🔗 Learn More
-- [Nature CRISPR Collection](https://www.nature.com/collections/crispr)
-- [CRISPR Clinical Trials Tracker](https://crisprtrials.org)
-
----
-
-*Last Updated: April 2025*
-`;
 	const [isCopied, setIsCopied] = useState(false);
 
 	const handleCopy = () => {
@@ -105,7 +39,7 @@ print(edit_gene("ATCGGATC", "GGA", "TTT"))  # Simulated gene edit
 
 	return (
 		<div className='relative w-full px-4 pb-8 pt-2 bg-transparent text-primary break-words whitespace-pre-wrap mb-2 group'>
-			<ReactMarkdown>{markdownText}</ReactMarkdown>
+			<ReactMarkdown>{text}</ReactMarkdown>
 
 			<button
 				className='relative mt-3 mb-2 rounded-xl'
