+++ frontend/src/Pages/404.jsx
@@ -35,12 +35,12 @@ export default function PageNotFound() {
 			className='grid h-full place-items-center bg-white px-6 py-24 sm:py-32 lg:px-8 flex flex-col h-screen overflow-hidden transition duration-300'
 			style={{ background: gradient }}>
 			<div className='text-center'>
-				<p className='text-xl font-semibold text-primary'>404</p>
-				<h1 className='mt-4 text-5xl font-semibold tracking-tight text-balance text-gray-900 sm:text-7xl'>
+				<p className='text-xl font-semibold text-accent'>404</p>
+				<h1 className='mt-4 text-5xl font-semibold tracking-tight text-balance text-primary sm:text-7xl'>
 					Page not found
 				</h1>
-				<p className='mt-6 text-lg font-medium text-pretty text-gray-500 sm:text-xl/8'>
-					Sorry, we couldn't find the page you're looking for.
+				<p className='mt-6 text-lg text-pretty text-primary sm:text-xl/8'>
+					Sorry, we couldn&apos;t find the page you&apos;re looking for.
 				</p>
 				<div className='mt-10 flex items-center justify-center gap-x-6'>
 					<Link
